Kernel malloc zone is allocated dynamically in contiguous physical memory.
The kernel malloc zone is allocated in contiguous physical memory in the
kernel higher half, past the bss section. Pages for this are mapped on
demand when a page fault occurs.

Began stubbing out a page frame allocator. Though, this is not integrated
yet with the page fault handler which services the kernel heap.