Extract kernel heap allocation to new class kernel_heap_allocator.
This classes allocates memory from a physically contiguous region of memory
which immediately follows the kernel image in memory. As pages are
allocated, they are mapped into virtual memory in the higher half at the
corresponding logical addresses. This is a change from before, where pages
are mapped on the next page fault at that address.

The allocator follows the memory_allocator interface. Though, free does
nothing at all and realloc has not yet been implemented.