Fix crash when trying to call global ctors on x86_64.
The page tables in the Kernel instance were too large for the mapped region
of memory. Kernel was big enough that the instance of g_kernel pushed the
instance of g_terminal out past the 2MB mark in the kernel image. As the
x86_64 bootstrap code only maps 2MB of memory in the higher half (this is
exactly one page table) this means that the kernel image is not entirely
mapped. Accesses to g_terminal result in a page fault.