Unit tests are split into a 32-bit test suite and a 64-bit test suite.
This fixes problems where code written to expect a 32-bit environment
could not run in a 64-bit test runner, and vice-versa. For example, it is
not possible to stash a pointer in an i386 PDE and follow it to the page
table when the test suite is running in a 64-bit environment.