Changes to make i386 and x86_64 versions of operation more similar.
The KernelAddressSpaceBootstrapper class's two versions can be made very
similar with small changes to PagingResolver. The logic to enumerate PDEs
in a given range of memory has been extracted to a new method on
PagingResolver called enumerate_page_directory_entries(). Additionally, the
method to fetch the page directory at a given linear address has been
replicated for the i386 version of the class; though, it doesn't use the
parameter for anything.