Encapsulate access to KERNEL_VIRTUAL_START_ADDR through the MMU class.
Try to (almost) never get the value of KERNEL_VIRTUAL_START_ADDR directly.
Instead, ask the MMU object, which is responsible for translations between
physical and linear addresses.

This gets us one step closer to being able to build the i386 paging code
as a x86_64 binary and test it in a x86_64 test runner.