The scheduler is able to schedule the initial kernel thread too.
The initial kernel thread is the thread with a stack allocated at boot time
and which initializes the kernel and scheduler in the first place. This
thread is now exposed to the scheduler and participates in the round-robin
scheduling queue.

NOTE: thread scheduling is still cooperative and requires a thread to call
      yield() to pass control to the next one.