Fix slow initialization of the physical memory map.
As an optimization, use an unmanaged physical memory map for the kernel
region of the address space and a managed memory map for the user portion.
This avoids allocating control structures for page tables and page frames
in kernel memory which will never change. The performance and memory
savings are very substantial.