Build with ninja, if available.
Build with the ninja CMake generator, if the ninja tool is available.
Building with ninja reduces the time for a full rebuild from 13s to 9s on
my machine.

Also, Ninja's output is filtered so that "error:" remains highlighted and
to resolve an issue where Ninja would display relative filenames like
"../../Common/foo" and VSCode's terminal was unable to link them.