Compile with -fcheck-new to avoid crashes on OOM.
By passing -fcheck-new, we instruct the compiler to check that new returned
a non-null pointer before attempting to construct the object. Since we also
build with -fno-exceptions, returning a null pointer is the only way to
signal failure.

Note that new(std::nothrow) is not an option because std::nothrow is not
included in the mandatory freestanding environment.