Malloc uses InterruptLock by default instead of Mutex.
The kernel must be careful about how and when it uses mutex locks which
interact with the scheduler. Using an InterruptLock here permits us to
use this allocator before the scheduler has been initialized, from
within the scheduler, and from interrupt context.

Though, these may not be good ideas. It merely makes it possible.